backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic[email]==2.7.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
pdfplumber==0.11.0
pymupdf==1.24.3
openai==1.30.1
anthropic==0.26.1
python-dotenv==1.0.1
aiofiles==23.2.1
email-validator==2.1.1
httpx<0.28
